Revise minified identifier, attribution
